+++ src/main/kotlin/ai/bigbear/pymatic/asksage/api/HttpRequestExecutor.kt
@@ -1,8 +1,0 @@
-package ai.bigbear.pymatic.asksage.api
-
-import java.net.http.HttpRequest
-import java.net.http.HttpResponse
-
-fun interface HttpRequestExecutor {
-    fun execute(request: HttpRequest): HttpResponse<String>
-}